refactors spline errors to separate module
